requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==2.3.2
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.4
flask-sqlalchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
MarkupSafe==2.1.5
numpy==1.24.4
packaging==25.0
pandas==2.0.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlalchemy==2.0.40
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
